Improvements: Some adaptation for gcc-10
Support windows compile with MSYS (partial). This part including:

- CMakeLists.txt using CMAKE_SHARED_LIBRARY_SUFFIX instead .so
- Replace finite() with std::isfinite()
  ref: https://github.com/facebookresearch/faiss/pull/1341
- Replace bzero (deprecated in POSIX 2001) with memset.
  ref: https://github.com/facebookresearch/faiss/pull/1340
- Some int types match.

+++ internal/core/src/index/knowhere/knowhere/index/vector_index/ConfAdapter.cpp
@@ -96,7 +96,7 @@ MatchNlist(int64_t size, int64_t nlist) {
 
     if (nlist * MIN_POINTS_PER_CENTROID > size) {
         // nlist is too large, adjust to a proper value
-        nlist = std::max(1L, size / MIN_POINTS_PER_CENTROID);
+        nlist = std::max(static_cast<int64_t>(1), size / MIN_POINTS_PER_CENTROID);
         LOG_KNOWHERE_WARNING_ << "Row num " << size << " match nlist " << nlist;
     }
     return nlist;
